AIML_Project/core/api_solutions/FAST API Solution/Phase1 - Core SQL API/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Oracle Database
cx_Oracle==8.3.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# CyberArk Integration
requests==2.31.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging & Monitoring
structlog==24.1.0